chore(deps): swap to django-storages-redux, an py3 compatible fork
This is a bit silly (LOC-wise) but I saw #161, and since I've been doing
this in some of my own projects, and thought this might be worth sharing.

[django-storages](https://pypi.python.org/pypi/django-storages) hasn't been
updated in over 2 years at this point.  Around December a work
[django-storages-redux](https://pypi.python.org/pypi/django-storages-redux/1.2.3)
was started to fix various outstanding issues as well as to make it Python 3 compatible.