feat(docs): add support for UML: sequence diagram
Adds support for js based rendered sequence diagram to explain the
server stacks and how/when they interact with each other.

This eliminates the use of any third party application to generate the
diagram and it's much faster to write and update.

__How to use:__
In markdown, write a code block with language as `sequence` following
the syntax from http://bramp.github.io/js-sequence-diagrams.

__Example__

<pre>
```sequence
a->b: A interacts with B
b->a: B responds
```
</pre>

__References__:
- http://en.wikipedia.org/wiki/Sequence_diagram
- http://bramp.github.io/js-sequence-diagrams

closes #48