refractor(apps): remove pages app
> Why was this change necessary?

This makes sure we are not enforcing any template layout structure. Frontend team
like to keep things at a single place. This is also reduce the number of folders
present in the project.

> How does it address the problem?

We moved the `pages.urls` to `urls` module, and templates are moved from
`pages/templates/pages` to `templates/pages`.

> Are there any side effects?

Nope, Flat is better than nested