refractor(tests): use django-dynamic-fixtures instead of factory-boy
* Why was this change necessary?

Simplifies writing factories and prevents duplication of model defination.

* How does it address the problem?

django-dynamic-fixtures provies two methods `G/get` and `N/new` that provides defaults
for almost all the model field types and default rules for custom field types can
be easily setup.

* Are there any side effects?

Not for any new projects, just that devs need to be familiar with DDF to be utilize
it at best! ;)